https://github.com/jackdewinter/pymarkdown/issues/582 - refactored leaf block processor
