+++ publish/coverage.json
@@ -6,8 +6,8 @@
         "totalCovered": 3581
     },
     "lineLevel": {
-        "totalMeasured": 14208,
-        "totalCovered": 14208
+        "totalMeasured": 14297,
+        "totalCovered": 14297
     }
 }
