+++ publish/pylint_suppression.json
@@ -87,13 +87,31 @@
         "pymarkdown/inline_response.py": {
             "too-many-instance-attributes": 1
         },
-        "pymarkdown/leaf_block_processor.py": {
-            "too-many-arguments": 21,
+        "pymarkdown/leaf_blocks/atx_leaf_block_processor.py": {
+            "too-many-arguments": 4
+        },
+        "pymarkdown/leaf_blocks/fenced_leaf_block_processor.py": {
+            "too-many-arguments": 9,
             "too-many-locals": 1
         },
-        "pymarkdown/leaf_block_processor_paragraph.py": {
+        "pymarkdown/leaf_blocks/indented_leaf_block_processor.py": {
+            "too-few-public-methods": 1,
+            "too-many-arguments": 4
+        },
+        "pymarkdown/leaf_blocks/leaf_block_helper.py": {
             "too-many-arguments": 1
         },
+        "pymarkdown/leaf_blocks/leaf_block_processor.py": {
+            "too-many-arguments": 1
+        },
+        "pymarkdown/leaf_blocks/leaf_block_processor_paragraph.py": {
+            "too-many-arguments": 1
+        },
+        "pymarkdown/leaf_blocks/setext_leaf_block_processor.py": {
+            "too-few-public-methods": 1,
+            "too-many-arguments": 2
+        },
+        "pymarkdown/leaf_blocks/thematic_leaf_block_processor.py": {},
         "pymarkdown/leaf_markdown_token.py": {
             "too-many-arguments": 4,
             "too-many-instance-attributes": 1
@@ -247,7 +265,7 @@
     "disables-by-name": {
         "too-many-arguments": 172,
         "too-many-locals": 16,
-        "too-few-public-methods": 9,
+        "too-few-public-methods": 11,
         "chained-comparison": 1,
         "too-many-boolean-expressions": 1,
         "too-many-instance-attributes": 19,
