+++ pymarkdown/leaf_blocks/leaf_block_processor_paragraph.py
@@ -25,8 +25,6 @@ from pymarkdown.tab_helper import TabHelper
 
 POGGER = ParserLogger(logging.getLogger(__name__))
 
-# pylint: disable=too-many-lines
-
 
 class LeafBlockProcessorParagraph:
     """
